requirements.txt
wheel
psycopg2
postgres
gunicorn
amqp==2.3.2
backcall==0.1.0
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
coinmarketcap==5.0.3
decorator==4.3.0
Django==2.0.7
django-cors-headers==2.3.0
djangorestframework==3.8.2
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
kombu==4.2.1
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
requests-cache==0.4.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
